compile unit: stage refactoring
Separate collision, atomic phy., iteration start and particle init.
