+++ include/picongpu/_defaultParam.loader
@@ -59,8 +59,6 @@
 #include "picongpu/param/shadowgraphy.param"
 #include "picongpu/param/species.param"
 #include "picongpu/param/speciesDefinition.param"
-#include "picongpu/param/speciesInitialization.param"
-#include "picongpu/param/iterationStart.param"
 #include "picongpu/param/collision.param"
 #include "picongpu/param/fieldSolver.param"
 #include "picongpu/param/fieldBackground.param"
