+++ include/picongpu/_defaultUnitless.loader
@@ -38,7 +38,6 @@
 #include "picongpu/unitless/ionizer.unitless"
 #include "picongpu/unitless/speciesAttributes.unitless"
 #include "picongpu/unitless/speciesDefinition.unitless"
-#include "picongpu/unitless/speciesInitialization.unitless"
 #include "picongpu/unitless/fieldBackground.unitless"
 #include "picongpu/unitless/collision.unitless"
 #include "picongpu/unitless/fileOutput.unitless"
