+++ include/picongpu/particles/boundary/Absorbing.hpp
@@ -24,6 +24,7 @@
 #include "picongpu/algorithms/Velocity.hpp"
 #include "picongpu/fields/absorber/absorber.hpp"
 #include "picongpu/fields/absorber/pml/Pml.kernel"
+#include "picongpu/particles/Manipulate.hpp"
 #include "picongpu/particles/boundary/ApplyImpl.hpp"
 #include "picongpu/particles/boundary/Kind.hpp"
 #include "picongpu/particles/boundary/Parameters.hpp"
