+++ share/picongpu/tests/KHI_growthRate/include/picongpu/param/precision.param
@@ -30,6 +30,7 @@
 
 #pragma once
 
+#include <picongpu/simulation_types.hpp>
 
 namespace picongpu
 {
